+++ updater/assets/drifterWhSystems.ts
@@ -1,0 +1,22 @@
+export default {
+  J055520: {
+    secondaryName: "Sentinel",
+    whClass: 14,
+  },
+  J110145: {
+    secondaryName: "Barbican",
+    whClass: 15,
+  },
+  J164710: {
+    secondaryName: "Vidette",
+    whClass: 16,
+  },
+  J200727: {
+    secondaryName: "Conflux",
+    whClass: 17,
+  },
+  J174618: {
+    secondaryName: "Redoubt",
+    whClass: 18,
+  },
+};